+++ entities/margin_levels.go
@@ -2,6 +2,7 @@ package entities
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"time"
 
@@ -108,4 +109,33 @@ func (ml MarginLevels) ToRow() []interface{} {
 var MarginLevelsColumns = []string{
 	"account_id", "timestamp", "maintenance_margin",
 	"search_level", "initial_margin", "collateral_release_level", "vega_time",
+}
+
+type MarginCursor struct {
+	VegaTime  time.Time
+	AccountID int64
+}
+
+func (mc MarginCursor) String() string {
+	bs, err := json.Marshal(mc)
+	if err != nil {
+		return fmt.Sprintf(`{"vegaTime":"%s","accountID":%d}`, mc.VegaTime, mc.AccountID)
+	}
+	return string(bs)
+}
+
+func (ml MarginLevels) Cursor() *Cursor {
+	cursor := MarginCursor{
+		VegaTime:  ml.VegaTime,
+		AccountID: ml.AccountID,
+	}
+	return NewCursor(cursor.String())
+}
+
+func ParseMarginLevelCursor(cursor string) (time.Time, int64, error) {
+	var mc MarginCursor
+	if err := json.Unmarshal([]byte(cursor), &mc); err != nil {
+		return time.Time{}, 0, fmt.Errorf("parsing margin cursor: %w", err)
+	}
+	return mc.VegaTime, mc.AccountID, nil
 }