+++ gateway/graphql/mocks/trading_data_service_client_v2_mock.go
@@ -136,6 +136,26 @@ func (mr *MockTradingDataServiceClientV2MockRecorder) GetERC20MultiSigSignerRemo
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetERC20MultiSigSignerRemovedBundles", reflect.TypeOf((*MockTradingDataServiceClientV2)(nil).GetERC20MultiSigSignerRemovedBundles), varargs...)
 }
 
+// GetMarginLevels mocks base method.
+func (m *MockTradingDataServiceClientV2) GetMarginLevels(arg0 context.Context, arg1 *v2.GetMarginLevelsRequest, arg2 ...grpc.CallOption) (*v2.GetMarginLevelsResponse, error) {
+	m.ctrl.T.Helper()
+	varargs := []interface{}{arg0, arg1}
+	for _, a := range arg2 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "GetMarginLevels", varargs...)
+	ret0, _ := ret[0].(*v2.GetMarginLevelsResponse)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// GetMarginLevels indicates an expected call of GetMarginLevels.
+func (mr *MockTradingDataServiceClientV2MockRecorder) GetMarginLevels(arg0, arg1 interface{}, arg2 ...interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]interface{}{arg0, arg1}, arg2...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetMarginLevels", reflect.TypeOf((*MockTradingDataServiceClientV2)(nil).GetMarginLevels), varargs...)
+}
+
 // GetMarketDataHistoryByID mocks base method.
 func (m *MockTradingDataServiceClientV2) GetMarketDataHistoryByID(arg0 context.Context, arg1 *v2.GetMarketDataHistoryByIDRequest, arg2 ...grpc.CallOption) (*v2.GetMarketDataHistoryByIDResponse, error) {
 	m.ctrl.T.Helper()