+++ sqlstore/margin_levels.go
@@ -3,6 +3,7 @@ package sqlstore
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"code.vegaprotocol.io/data-node/entities"
 	"code.vegaprotocol.io/data-node/metrics"
@@ -45,10 +46,24 @@ func (ml *MarginLevels) Flush(ctx context.Context) error {
 }
 
 func (ml *MarginLevels) GetMarginLevelsByID(ctx context.Context, partyID, marketID string, pagination entities.OffsetPagination) ([]entities.MarginLevels, error) {
+	whereClause, bindVars := buildAccountWhereClause(partyID, marketID)
+
+	query := fmt.Sprintf(`select distinct on (account_id) %s
+		from all_margin_levels
+		%s
+		order by account_id, vega_time desc`, sqlMarginLevelColumns,
+		whereClause)
+
+	query, bindVars = orderAndPaginateQuery(query, nil, pagination, bindVars...)
+	defer metrics.StartSQLQuery("MarginLevels", "GetByID")()
+	var marginLevels []entities.MarginLevels
+	err := pgxscan.Select(ctx, ml.Connection, &marginLevels, query, bindVars...)
+	return marginLevels, err
+}
+
+func buildAccountWhereClause(partyID, marketID string) (string, []interface{}) {
 	party := entities.NewPartyID(partyID)
 	market := entities.NewMarketID(marketID)
-
-	var marginLevels []entities.MarginLevels
 
 	var bindVars []interface{}
 
@@ -72,17 +87,43 @@ func (ml *MarginLevels) GetMarginLevelsByID(ctx context.Context, partyID, market
 		accountsWhereClause = fmt.Sprintf("where %s", whereMarket)
 	}
 
-	whereClause := fmt.Sprintf("where all_margin_levels.account_id  in (select id from accounts %s)", accountsWhereClause)
+	return fmt.Sprintf("where all_margin_levels.account_id  in (select id from accounts %s)", accountsWhereClause), bindVars
+}
+
+func (ml *MarginLevels) GetMarginLevelsByIDWithCursorPagination(ctx context.Context, partyID, marketID string, pagination entities.Pagination) ([]entities.MarginLevels, entities.PageInfo, error) {
+	whereClause, bindVars := buildAccountWhereClause(partyID, marketID)
 
 	query := fmt.Sprintf(`select distinct on (account_id) %s
 		from all_margin_levels
-		%s
-		order by account_id, vega_time desc`, sqlMarginLevelColumns,
+		%s`, sqlMarginLevelColumns,
 		whereClause)
 
-	query, bindVars = orderAndPaginateQuery(query, nil, pagination, bindVars...)
-	defer metrics.StartSQLQuery("MarginLevels", "GetByID")()
-	err := pgxscan.Select(ctx, ml.Connection, &marginLevels, query, bindVars...)
+	sorting, cmp, cursor := extractPaginationInfo(pagination)
+	var (
+		vegaTime  time.Time
+		accountID int64
+		err       error
+	)
 
-	return marginLevels, err
+	if cursor != "" {
+		vegaTime, accountID, err = entities.ParseMarginLevelCursor(cursor)
+		if err != nil {
+			return nil, entities.PageInfo{}, fmt.Errorf("parsing cursor: %w", err)
+		}
+	}
+
+	builders := []CursorQueryParameter{
+		NewCursorQueryParameter("account_id", sorting, cmp, accountID),
+		NewCursorQueryParameter("vega_time", sorting, cmp, vegaTime),
+	}
+
+	query, bindVars = orderAndPaginateWithCursor(query, pagination, builders, bindVars...)
+	var marginLevels []entities.MarginLevels
+
+	if err := pgxscan.Select(ctx, ml.Connection, &marginLevels, query, bindVars...); err != nil {
+		return nil, entities.PageInfo{}, err
+	}
+
+	pagedMargins, pageInfo := entities.PageEntities(marginLevels, pagination)
+	return pagedMargins, pageInfo, nil
 }