+++ sqlstore/parties.go
@@ -105,7 +105,7 @@ func (ps *Parties) GetAllPaged(ctx context.Context, partyID string, pagination e
 	var pageInfo entities.PageInfo
 
 	sorting, cmp, cursor := extractPaginationInfo(pagination)
-	cursors := []CursorBuilder{NewCursorBuilder("vega_time", sorting, cmp, cursor)}
+	cursors := []CursorQueryParameter{NewCursorQueryParameter("vega_time", sorting, cmp, cursor)}
 	query, args = orderAndPaginateWithCursor(query, pagination, cursors, args...)
 
 	if err := pgxscan.Select(ctx, ps.Connection, &parties, query, args...); err != nil {